Completion of dashboard.
Dashboard now displays all anilox roller as well as have slight color coding to provide visual
aid for knowing when an anilox needs to be clean via the row being red or is close to needing
a cleaning via the row being yellow. These should be updated later to allow end-user to have more
control over the color pallet.