Load anilox list complete
Users can now load a list of anilox from a Comma Seperate Values (CSV) file. This file has to be formatted
with the anilox roller id in the first position, the LPI in the second position and the BCM in the third
position. If any data is missing, the user will be made aware as well as if any issues occur during the import
process via a dialog box with some details.