Completed adding new anilox rollers.
Users can now add new anilox rollers from the settings menu. Window closure handling is also prevelent in
the top level create as well as the base application so if a user closes the application, the top level
window will also be closed. The buttons in the setting menu are disabled to prevent multiple instances of
the top level window from being created. Small changes made to the main applications:
- Disable resizing
- Creation of close function (to handle destroying top level windows)
- Creation of style function so future style changes can be cohesive.
- Moved over testing color pallet. This will be replace with json values so a custom color pallet can be
	created by the end user